chore: 클래스 이름 변경
- logBat -> LogBat 클래스 이름을 변경했습니다.
- Java-Doc에 저자를 추가했습니다.
