+++ sdk/java/logbat-sdk/src/main/java/info/logbat/adapter/logback/LogBatAppender.java
@@ -3,11 +3,11 @@ package info.logbat.adapter.logback;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.UnsynchronizedAppenderBase;
 import info.logbat.LogBatFactory;
-import info.logbat.application.Logbat;
+import info.logbat.application.LogBat;
 
 public class LogBatAppender extends UnsynchronizedAppenderBase<ILoggingEvent> {
 
-    private Logbat logbat;
+    private LogBat logbat;
 
     @Override
     public void start() {
